iterate, clean up
