+++ src/components/Explorer.js
@@ -6,7 +6,7 @@ import ExplorerHeader from './ExplorerHeader'
 import NibrsContainer from './NibrsContainer'
 import NotFound from './NotFound'
 import Sidebar from './Sidebar'
-import Sparklines from './Sparklines'
+import SparklineSection from './SparklineSection'
 import TrendContainer from './TrendContainer'
 import UcrParticipationInformation from './UcrParticipationInformation'
 import { updateApp } from '../actions/composite'
@@ -123,7 +123,7 @@ class Explorer extends React.Component {
             />
             <hr className="mt0 mb3" />
             {isAgency &&
-              <Sparklines
+              <SparklineSection
                 crime={crime}
                 place={place}
                 since={since}
